refactor: generateJsonFiles
